+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Fix #1069: Setting up automated renewal of TLS certificate for productions
+
 ## v4.4.13 - 2025-06-13 - 26083397b - live since 2025-06-16 
 
 - Feat #199: Reorder home page sections
@@ -46,6 +48,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## v4.4.8 - 2025-04-08 - 01de0477b - not deployed to live
 
+- Fix #1069: Setting up automated renewal of TLS certificates for productions
 - Fix #2033: Create a mockup for all upload statuses except published
 
 ## v4.4.7 - 2025-03-26 - 4c548f037 - live since 2025-04-03